+++ packages/vseed/src/types/chartType/areaPercent/zAreaPercent.ts
@@ -32,7 +32,7 @@ export const zAreaPercent = z.object({
   measures: zMeasureTree.nullish(),
   backgroundColor: zBackgroundColor.nullish(),
   color: zColor.nullish(),
-  label: zLabel.nullish(),
+  label: zLabel.omit({ labelLayout: true }).nullish(),
   legend: zLegend.nullish(),
   tooltip: zTooltip.nullish(),
   xAxis: zXBandAxis.nullish(),