+++ packages/vseed/src/types/chartType/heatmap/heatmap.ts
@@ -82,7 +82,7 @@ export interface Heatmap {
   /**
    * @description 热力图标签配置, 用于定义图表的数据标签, 自动开启标签反色, 确保标签可读性.
    */
-  label?: Label
+  label?: Omit<Label, 'labelLayout'>
 
   /**
    * 图例