+++ packages/vseed/src/types/chartType/heatmap/zHeatmap.ts
@@ -20,7 +20,7 @@ export const zHeatmap = z.object({
   measures: zMeasureTree.nullish(),
   backgroundColor: zBackgroundColor.nullish(),
   color: zColor.nullish(),
-  label: zLabel.nullish(),
+  label: zLabel.omit({ labelLayout: true }).nullish(),
   legend: zColorLegend.nullish(),
   tooltip: zTooltip.nullish(),
   theme: zTheme.nullish(),