+++ packages/vseed/src/types/chartType/roseParallel/zRoseParallel.ts
@@ -21,7 +21,7 @@ export const zRoseParallel = z.object({
   measures: zMeasureTree.nullish(),
   backgroundColor: zBackgroundColor.nullish(),
   color: zColor.nullish(),
-  label: zLabel.nullish(),
+  label: zLabel.omit({ labelLayout: true }).nullish(),
   legend: zLegend.nullish(),
   tooltip: zTooltip.nullish(),
   theme: zTheme.nullish(),